+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - [1102](https://github.com/vegaprotocol/core-test-coverage/issues/1102) - Add coverage for `0042-LIQF-084`
 - [11121](https://github.com/vegaprotocol/vega/issues/11121) - Remove auction trigger staleness functionality
 - [11127](https://github.com/vegaprotocol/vega/issues/11127) - Price monitoring engine should record all observations with the same weight
+- [10995](https://github.com/vegaprotocol/vega/issues/10995) - Liquidation range defined by its own parameter.
 
 ### 🐛 Fixes